+++ weblate/checks/tests/test_chars_checks.py
@@ -20,6 +20,7 @@ from weblate.checks.chars import (
     EndSpaceCheck,
     EndStopCheck,
     EscapedNewlineCountingCheck,
+    KabyleCharactersCheck,
     KashidaCheck,
     MaxLengthCheck,
     NewLineCountCheck,
@@ -476,4 +477,27 @@ class PunctuationSpacingCheckTest(CheckTestCase):
                 "rst-text",
             ),
             "fr",
+        )
+
+
+class KabyleCharactersCheckTest(CheckTestCase):
+    check = KabyleCharactersCheck()
+    default_lang = "kab"
+
+    def setUp(self) -> None:
+        super().setUp()
+        self.test_good_matching = ("string", "string", "")
+        self.test_failure_1 = ("string", "γ", "")
+        self.test_failure_2 = ("string", "Γ", "")
+        self.test_failure_3 = ("string", "ε", "")
+
+    def test_skip(self) -> None:
+        self.do_test(
+            False,
+            (
+                "string",
+                "γΓε",
+                "",
+            ),
+            "el",
         )